pyproject.toml
[build-system]
requires = ["setuptools>=69.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "edm_store"
version = "0.1.0"
description = "EDM-Store is a distributed remote sensing image management tool written in python."
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
authors = []
maintainers = []
dependencies = [
  "affine~=2.4.0",
  "awsauth~=0.3.3",
  "boto~=2.49.0",
  "numpy~=1.26.3",
  "pydantic~=2.5.3",
  "pymongo~=4.6.1",
  "pyproj~=3.6.1",
  "rasterio~=1.3.9",
  "requests~=2.31.0",
  "requests-aws~=0.1.8",
  "shapely~=2.0.2",
  "pyyaml~=6.0.1"
]

[project.scripts]
edm_store = "edm_store.scripts.app:cli"
